HTTP周りを別のclassに実装
